refactor(BackBreeze.Scrollbar): move scrollbar functions into own module
The scrollbar structs and functionality now all exist inside of
`BackBreeze.Scrollbar`

Some additional tests and fixes have been added around the extremes for
the tops and bottoms of the scrollbars.
