+++ lib/back_breeze/box.ex
@@ -6,52 +6,6 @@ defmodule BackBreeze.Box do
   """
   alias BackBreeze.Ucwidth
 
-  defmodule ScrollbarContext do
-    @moduledoc false
-
-    defstruct style: %BackBreeze.Style{},
-              config: %BackBreeze.Scrollbar{},
-              scroll: {0, 0},
-              metrics: %{},
-              max_x: 0,
-              max_y: 0,
-              left: 0,
-              top: 0,
-              right: 0,
-              bottom: 0,
-              viewport_width: 0,
-              viewport_height: 0,
-              content_width: 0,
-              content_height: 0,
-              vertical?: false,
-              horizontal?: false,
-              eff_viewport_width: 0,
-              eff_viewport_height: 0,
-              vertical_placement: :end,
-              horizontal_placement: :end,
-              x_scrollbar: 0,
-              y_scrollbar: 0,
-              vertical_start: 0,
-              vertical_end: -1,
-              horizontal_start: 0,
-              horizontal_end: -1,
-              scroll_top: 0,
-              scroll_left: 0
-  end
-
-  defmodule AxisContext do
-    @moduledoc false
-
-    defstruct axis: :vertical,
-              config: %BackBreeze.Scrollbar{},
-              fixed: 0,
-              start: 0,
-              stop: -1,
-              scroll_value: 0,
-              content_size: 0,
-              viewport_size: 0
-  end
-
   defstruct content: "",
             children: [],
             style: %BackBreeze.Style{},
@@ -59,10 +13,10 @@ defmodule BackBreeze.Box do
             height: nil,
             state: :ready,
             position: :relative,
-            display: :block,
             left: nil,
-            scroll: {0, 0},
             top: nil,
+            display: :block,
+            scroll: {0, 0},
             layer: 0,
             layer_map: %{}
 
@@ -88,6 +42,7 @@ defmodule BackBreeze.Box do
       end
 
     style = struct(BackBreeze.Style, style)
+
     struct(BackBreeze.Box, Map.put(map, :style, style))
   end
 
@@ -137,13 +92,11 @@ defmodule BackBreeze.Box do
         {layer_map, max_width, max_height} = generate_layer_map(content, %{}, 0, 0)
 
         layer_map =
-          maybe_add_scrollbars(layer_map, %{
+          BackBreeze.Scrollbar.add_to_layer_map(layer_map, %{
             style: box.style,
             scroll: box.scroll,
-            metrics: %{
-              content_height: dimensions.content_height,
-              content_width: raw_content_width(box.content)
-            },
+            content_height: dimensions.content_height,
+            content_width: raw_content_width(box.content),
             max_x: max_width,
             max_y: max_height
           })
@@ -238,17 +191,18 @@ defmodule BackBreeze.Box do
       end
 
     child_layer_map =
-      maybe_add_scrollbars(child_layer_map, %{
+      BackBreeze.Scrollbar.add_to_layer_map(child_layer_map, %{
         style: box.style,
         scroll: box.scroll,
-        metrics: %{content_height: dimensions.content_height, content_width: child_width},
+        content_height: dimensions.content_height,
+        content_width: child_width,
         max_x: max_width,
         max_y: max_height
       })
 
     {start_x, start_y} =
-      case box do
-        %{position: :absolute, left: left, top: top} -> {left, top}
+      case box.position do
+        :absolute -> {box.left, box.top}
         _ -> {0, 0}
       end
 
@@ -490,479 +444,6 @@ defmodule BackBreeze.Box do
       end)
 
     Enum.join(content, "\n") |> String.trim_trailing("\n")
-  end
-
-  defp maybe_add_scrollbars(layer_map, options)
-       when is_map(options) and not is_struct(options, ScrollbarContext) do
-    options
-    |> build_scrollbar_context()
-    |> then(&maybe_add_scrollbars(layer_map, &1))
-  end
-
-  defp maybe_add_scrollbars(layer_map, %ScrollbarContext{} = context) do
-    cond do
-      context.style.overflow != :hidden or !context.config.enabled ->
-        layer_map
-
-      !is_integer(context.max_x) or !is_integer(context.max_y) ->
-        layer_map
-
-      true ->
-        context = prepare_scrollbar_context(context)
-
-        if !context.vertical? and !context.horizontal? do
-          layer_map
-        else
-          layer_map
-          |> maybe_clear_scrollbar_strips(context)
-          |> maybe_draw_vertical_scrollbar(context)
-          |> maybe_draw_horizontal_scrollbar(context)
-          |> maybe_draw_intersection(context)
-        end
-    end
-  end
-
-  defp build_scrollbar_context(%{style: style} = options) do
-    %ScrollbarContext{
-      style: style,
-      config: BackBreeze.Scrollbar.normalize(style.scrollbar, style),
-      scroll: Map.get(options, :scroll, {0, 0}),
-      metrics: Map.get(options, :metrics, %{}),
-      max_x: Map.get(options, :max_x),
-      max_y: Map.get(options, :max_y)
-    }
-  end
-
-  defp prepare_scrollbar_context(%ScrollbarContext{} = context) do
-    {left, top, right, bottom} =
-      viewport_bounds(context.style.border, context.max_x, context.max_y)
-
-    viewport_width = max(right - left + 1, 0)
-    viewport_height = max(bottom - top + 1, 0)
-
-    metrics = if is_map(context.metrics), do: context.metrics, else: %{}
-
-    content_height = max(Map.get(metrics, :content_height, viewport_height), viewport_height)
-    content_width = max(Map.get(metrics, :content_width, viewport_width), viewport_width)
-
-    %{
-      context
-      | left: left,
-        top: top,
-        right: right,
-        bottom: bottom,
-        viewport_width: viewport_width,
-        viewport_height: viewport_height,
-        content_width: content_width,
-        content_height: content_height
-    }
-    |> resolve_visible_axes()
-    |> set_scrollbar_positions()
-    |> set_scroll_offsets()
-  end
-
-  defp set_scrollbar_positions(%ScrollbarContext{} = context) do
-    vertical_placement = BackBreeze.Scrollbar.placement(context.config, :vertical)
-    horizontal_placement = BackBreeze.Scrollbar.placement(context.config, :horizontal)
-
-    {x_scrollbar, y_scrollbar} =
-      scrollbar_positions(context, vertical_placement, horizontal_placement)
-
-    {vertical_start, vertical_end} =
-      trim_axis_for_intersection(
-        context.top,
-        context.bottom,
-        context.horizontal?,
-        horizontal_placement
-      )
-
-    {horizontal_start, horizontal_end} =
-      trim_axis_for_intersection(
-        context.left,
-        context.right,
-        context.vertical?,
-        vertical_placement
-      )
-
-    %{
-      context
-      | vertical_placement: vertical_placement,
-        horizontal_placement: horizontal_placement,
-        x_scrollbar: x_scrollbar,
-        y_scrollbar: y_scrollbar,
-        vertical_start: vertical_start,
-        vertical_end: vertical_end,
-        horizontal_start: horizontal_start,
-        horizontal_end: horizontal_end
-    }
-  end
-
-  defp set_scroll_offsets(%ScrollbarContext{} = context) do
-    {scroll_top, scroll_left} = normalize_scroll(context.scroll)
-    %{context | scroll_top: scroll_top, scroll_left: scroll_left}
-  end
-
-  defp maybe_clear_scrollbar_strips(layer_map, context) do
-    layer_map
-    |> maybe_clear_vertical_scrollbar_strip(context)
-    |> maybe_clear_horizontal_scrollbar_strip(context)
-  end
-
-  defp maybe_clear_vertical_scrollbar_strip(
-         layer_map,
-         %ScrollbarContext{
-           config: %{mode: :inset},
-           vertical?: true,
-           x_scrollbar: x_scrollbar,
-           left: left,
-           right: right,
-           vertical_start: vertical_start,
-           vertical_end: vertical_end
-         }
-       )
-       when x_scrollbar >= left and x_scrollbar <= right do
-    clear_vertical_strip(layer_map, x_scrollbar, vertical_start, vertical_end)
-  end
-
-  defp maybe_clear_vertical_scrollbar_strip(layer_map, _context), do: layer_map
-
-  defp maybe_clear_horizontal_scrollbar_strip(
-         layer_map,
-         %ScrollbarContext{
-           config: %{mode: :inset},
-           horizontal?: true,
-           y_scrollbar: y_scrollbar,
-           top: top,
-           bottom: bottom,
-           horizontal_start: horizontal_start,
-           horizontal_end: horizontal_end
-         }
-       )
-       when y_scrollbar >= top and y_scrollbar <= bottom do
-    clear_horizontal_strip(layer_map, y_scrollbar, horizontal_start, horizontal_end)
-  end
-
-  defp maybe_clear_horizontal_scrollbar_strip(layer_map, _context), do: layer_map
-
-  defp viewport_bounds(border, max_x, max_y) do
-    left = if border.left, do: 1, else: 0
-    top = if border.top, do: 1, else: 0
-
-    right = max(max_x - if(border.right, do: 1, else: 0), left)
-    bottom = max(max_y - if(border.bottom, do: 1, else: 0), top)
-
-    {left, top, right, bottom}
-  end
-
-  defp scrollbar_positions(
-         %ScrollbarContext{
-           style: %{border: border},
-           config: config,
-           left: left,
-           top: top,
-           right: right,
-           bottom: bottom
-         },
-         vertical_placement,
-         horizontal_placement
-       ) do
-    x_scrollbar =
-      cond do
-        config.mode == :inset and vertical_placement == :start and border.left -> left - 1
-        config.mode == :inset and vertical_placement == :end and border.right -> right + 1
-        vertical_placement == :start -> left
-        true -> right
-      end
-
-    y_scrollbar =
-      cond do
-        config.mode == :inset and horizontal_placement == :start and border.top -> top - 1
-        config.mode == :inset and horizontal_placement == :end and border.bottom -> bottom + 1
-        horizontal_placement == :start -> top
-        true -> bottom
-      end
-
-    {x_scrollbar, y_scrollbar}
-  end
-
-  defp resolve_visible_axes(%ScrollbarContext{} = context) do
-    config = context.config
-
-    vertical? =
-      BackBreeze.Scrollbar.axis_enabled?(config, :vertical) and
-        BackBreeze.Scrollbar.visible?(
-          config,
-          :vertical,
-          context.content_height,
-          context.viewport_height
-        )
-
-    horizontal? =
-      BackBreeze.Scrollbar.axis_enabled?(config, :horizontal) and
-        BackBreeze.Scrollbar.visible?(
-          config,
-          :horizontal,
-          context.content_width,
-          context.viewport_width
-        )
-
-    {eff_viewport_width, eff_viewport_height} =
-      BackBreeze.Scrollbar.effective_viewport_size(config, %{
-        width: context.viewport_width,
-        height: context.viewport_height,
-        vertical?: vertical?,
-        horizontal?: horizontal?
-      })
-
-    vertical? =
-      BackBreeze.Scrollbar.axis_enabled?(config, :vertical) and
-        BackBreeze.Scrollbar.visible?(
-          config,
-          :vertical,
-          context.content_height,
-          eff_viewport_height
-        )
-
-    horizontal? =
-      BackBreeze.Scrollbar.axis_enabled?(config, :horizontal) and
-        BackBreeze.Scrollbar.visible?(
-          config,
-          :horizontal,
-          context.content_width,
-          eff_viewport_width
-        )
-
-    {eff_viewport_width, eff_viewport_height} =
-      BackBreeze.Scrollbar.effective_viewport_size(config, %{
-        width: context.viewport_width,
-        height: context.viewport_height,
-        vertical?: vertical?,
-        horizontal?: horizontal?
-      })
-
-    %{
-      context
-      | vertical?: vertical?,
-        horizontal?: horizontal?,
-        eff_viewport_width: eff_viewport_width,
-        eff_viewport_height: eff_viewport_height
-    }
-  end
-
-  defp normalize_scroll({top, left}) do
-    top = if is_integer(top), do: max(top, 0), else: 0
-    left = if is_integer(left), do: max(left, 0), else: 0
-    {top, left}
-  end
-
-  defp normalize_scroll(_), do: {0, 0}
-
-  defp trim_axis_for_intersection(start_pos, end_pos, other_enabled?, other_placement) do
-    start_pos = if other_enabled? && other_placement == :start, do: start_pos + 1, else: start_pos
-    end_pos = if other_enabled? && other_placement == :end, do: end_pos - 1, else: end_pos
-
-    {start_pos, max(start_pos - 1, end_pos)}
-  end
-
-  defp clear_vertical_strip(layer_map, x, y_start, y_end) when y_start <= y_end do
-    Enum.reduce(y_start..y_end, layer_map, fn y, acc ->
-      Map.put(acc, {y, x}, {" ", ""})
-    end)
-  end
-
-  defp clear_vertical_strip(layer_map, _x, _y_start, _y_end), do: layer_map
-
-  defp clear_horizontal_strip(layer_map, y, x_start, x_end) when x_start <= x_end do
-    Enum.reduce(x_start..x_end, layer_map, fn x, acc ->
-      Map.put(acc, {y, x}, {" ", ""})
-    end)
-  end
-
-  defp clear_horizontal_strip(layer_map, _y, _x_start, _x_end), do: layer_map
-
-  defp maybe_draw_vertical_scrollbar(layer_map, %ScrollbarContext{vertical?: true} = context) do
-    context
-    |> vertical_axis_context()
-    |> then(&draw_axis_scrollbar(layer_map, &1))
-  end
-
-  defp maybe_draw_vertical_scrollbar(layer_map, _context), do: layer_map
-
-  defp maybe_draw_horizontal_scrollbar(
-         layer_map,
-         %ScrollbarContext{horizontal?: true} = context
-       ) do
-    context
-    |> horizontal_axis_context()
-    |> then(&draw_axis_scrollbar(layer_map, &1))
-  end
-
-  defp maybe_draw_horizontal_scrollbar(layer_map, _context), do: layer_map
-
-  defp vertical_axis_context(%ScrollbarContext{} = context) do
-    %AxisContext{
-      axis: :vertical,
-      config: context.config,
-      fixed: context.x_scrollbar,
-      start: context.vertical_start,
-      stop: context.vertical_end,
-      scroll_value: context.scroll_top,
-      content_size: context.content_height,
-      viewport_size: context.eff_viewport_height
-    }
-  end
-
-  defp horizontal_axis_context(%ScrollbarContext{} = context) do
-    %AxisContext{
-      axis: :horizontal,
-      config: context.config,
-      fixed: context.y_scrollbar,
-      start: context.horizontal_start,
-      stop: context.horizontal_end,
-      scroll_value: context.scroll_left,
-      content_size: context.content_width,
-      viewport_size: context.eff_viewport_width
-    }
-  end
-
-  defp draw_axis_scrollbar(layer_map, %AxisContext{} = axis_context) do
-    total_size = axis_context.stop - axis_context.start + 1
-
-    if total_size <= 0 do
-      layer_map
-    else
-      {track_start, track_end, layer_map} =
-        maybe_draw_axis_arrows(layer_map, axis_context, total_size)
-
-      draw_scroll_track_and_thumb(layer_map, %{
-        axis_context
-        | start: track_start,
-          stop: track_end
-      })
-    end
-  end
-
-  defp maybe_draw_axis_arrows(
-         layer_map,
-         %AxisContext{config: %{arrows: true}} = axis_context,
-         total_size
-       )
-       when total_size >= 3 do
-    layer_map =
-      put_axis_segment(
-        layer_map,
-        axis_context,
-        axis_context.start,
-        axis_arrow_segment(axis_context, :start)
-      )
-
-    layer_map =
-      put_axis_segment(
-        layer_map,
-        axis_context,
-        axis_context.stop,
-        axis_arrow_segment(axis_context, :end)
-      )
-
-    {axis_context.start + 1, axis_context.stop - 1, layer_map}
-  end
-
-  defp maybe_draw_axis_arrows(layer_map, %AxisContext{} = axis_context, _total_size) do
-    {axis_context.start, axis_context.stop, layer_map}
-  end
-
-  defp draw_scroll_track_and_thumb(
-         layer_map,
-         %AxisContext{start: track_start, stop: track_end, viewport_size: viewport_size} =
-           axis_context
-       )
-       when track_start <= track_end and viewport_size > 0 do
-    track_size = track_end - track_start + 1
-
-    {track_segment, thumb_segment} = axis_segments(axis_context)
-
-    max_scroll = max(axis_context.content_size - axis_context.viewport_size, 0)
-    scroll_value = min(max(axis_context.scroll_value, 0), max_scroll)
-
-    thumb_size =
-      BackBreeze.Scrollbar.thumb_size(
-        axis_context.config,
-        track_size,
-        max(axis_context.content_size, axis_context.viewport_size)
-      )
-
-    thumb_start =
-      if max_scroll == 0 or thumb_size >= track_size do
-        track_start
-      else
-        offset = round(scroll_value * (track_size - thumb_size) / max_scroll)
-        track_start + offset
-      end
-
-    thumb_end = min(thumb_start + thumb_size - 1, track_end)
-
-    layer_map =
-      Enum.reduce(track_start..track_end, layer_map, fn pos, acc ->
-        put_axis_segment(acc, axis_context, pos, track_segment)
-      end)
-
-    Enum.reduce(thumb_start..thumb_end, layer_map, fn pos, acc ->
-      put_axis_segment(acc, axis_context, pos, thumb_segment)
-    end)
-  end
-
-  defp draw_scroll_track_and_thumb(layer_map, _axis_context), do: layer_map
-
-  defp axis_segments(%AxisContext{axis: :vertical, config: config}) do
-    {config.vertical.track, config.vertical.thumb}
-  end
-
-  defp axis_segments(%AxisContext{axis: :horizontal, config: config}) do
-    {config.horizontal.track, config.horizontal.thumb}
-  end
-
-  defp axis_arrow_segment(%AxisContext{axis: :vertical, config: config}, :start),
-    do: config.vertical.arrow_start
-
-  defp axis_arrow_segment(%AxisContext{axis: :vertical, config: config}, :end),
-    do: config.vertical.arrow_end
-
-  defp axis_arrow_segment(%AxisContext{axis: :horizontal, config: config}, :start),
-    do: config.horizontal.arrow_start
-
-  defp axis_arrow_segment(%AxisContext{axis: :horizontal, config: config}, :end),
-    do: config.horizontal.arrow_end
-
-  defp put_axis_segment(layer_map, %AxisContext{} = axis_context, pos, segment) do
-    put_segment(layer_map, axis_point(axis_context, pos), segment)
-  end
-
-  defp axis_point(%AxisContext{axis: :vertical, fixed: fixed}, pos), do: {pos, fixed}
-  defp axis_point(%AxisContext{axis: :horizontal, fixed: fixed}, pos), do: {fixed, pos}
-
-  defp maybe_draw_intersection(
-         layer_map,
-         %ScrollbarContext{vertical?: true, horizontal?: true} = context
-       ) do
-    put_segment(
-      layer_map,
-      {context.y_scrollbar, context.x_scrollbar},
-      context.config.intersection
-    )
-  end
-
-  defp maybe_draw_intersection(layer_map, _context), do: layer_map
-
-  defp put_segment(layer_map, _point, nil), do: layer_map
-
-  defp put_segment(layer_map, point, segment) do
-    char = BackBreeze.Scrollbar.segment_char(segment)
-
-    if is_binary(char) do
-      Map.put(layer_map, point, {char, BackBreeze.Scrollbar.style_sequence(segment)})
-    else
-      layer_map
-    end
   end
 
   defp clip_child_layer_map(layer_map, %{overflow: :hidden, border: border}, max_x, max_y)
