+++ lib/back_breeze/scrollbar.ex
@@ -1,5 +1,51 @@
+defmodule BackBreeze.Scrollbar.Context do
+  @moduledoc false
+
+  defstruct style: %BackBreeze.Style{},
+            max_x: 0,
+            max_y: 0,
+            left: 0,
+            top: 0,
+            right: 0,
+            bottom: 0,
+            viewport_width: 0,
+            viewport_height: 0,
+            content_width: 0,
+            content_height: 0,
+            vertical?: false,
+            horizontal?: false,
+            eff_viewport_width: 0,
+            eff_viewport_height: 0,
+            vertical_placement: :end,
+            horizontal_placement: :end,
+            x_scrollbar: 0,
+            y_scrollbar: 0,
+            vertical_start: 0,
+            vertical_end: -1,
+            horizontal_start: 0,
+            horizontal_end: -1,
+            scroll_top: 0,
+            scroll_left: 0
+end
+
+defmodule BackBreeze.Scrollbar.AxisContext do
+  @moduledoc false
+
+  defstruct axis: :vertical,
+            config: nil,
+            fixed: 0,
+            start: 0,
+            stop: -1,
+            scroll_value: 0,
+            content_size: 0,
+            viewport_size: 0
+end
+
 defmodule BackBreeze.Scrollbar do
   @moduledoc false
+
+  alias BackBreeze.Scrollbar.Context
+  alias BackBreeze.Scrollbar.AxisContext
 
   defmodule Segment do
     @moduledoc false
@@ -11,16 +57,6 @@ defmodule BackBreeze.Scrollbar do
               italic: false,
               reverse: false,
               style: ""
-  end
-
-  defmodule AxisSegmentsContext do
-    @moduledoc false
-
-    defstruct base: %{},
-              common_thumb: %{},
-              common_track: %{},
-              overrides: %{},
-              fallback_color: nil
   end
 
   defstruct enabled: false,
@@ -40,9 +76,6 @@ defmodule BackBreeze.Scrollbar do
   @type t :: %__MODULE__{}
 
   @axes [:vertical, :horizontal, :both]
-  @shows [:auto, :always, :never, :focus]
-  @modes [:overlay, :inset]
-  @placements [:start, :end]
 
   @spec normalize(term(), map()) :: t()
   def normalize(value, style \\ %BackBreeze.Style{})
@@ -65,76 +98,42 @@ defmodule BackBreeze.Scrollbar do
 
   def normalize(map, style) when is_map(map) do
     base = default(style)
-
-    axis = normalize_axis(Map.get(map, :axis, base.axis))
-    show = normalize_show(Map.get(map, :show, base.show))
-    mode = normalize_mode(Map.get(map, :mode, base.mode))
-    placement = normalize_placement(Map.get(map, :placement, base.placement))
-
-    min_thumb_size =
-      map
-      |> Map.get(:min_thumb_size, base.min_thumb_size)
-      |> normalize_int(base.min_thumb_size)
-      |> max(1)
-
-    sizing = normalize_sizing(Map.get(map, :sizing, base.sizing))
-    arrows = normalize_boolean(Map.get(map, :arrows, base.arrows), base.arrows)
-
-    vertical_placement =
-      map
-      |> Map.get(:vertical_placement)
-      |> normalize_optional_placement()
-
-    horizontal_placement =
-      map
-      |> Map.get(:horizontal_placement)
-      |> normalize_optional_placement()
-
     fallback_color = Map.get(style, :border_color)
     common_thumb = Map.get(map, :thumb, %{})
     common_track = Map.get(map, :track, %{})
 
-    vertical_overrides = Map.get(map, :vertical, %{})
-    horizontal_overrides = Map.get(map, :horizontal, %{})
-
     vertical =
-      normalize_axis_segments(%AxisSegmentsContext{
-        base: base.vertical,
-        common_thumb: common_thumb,
-        common_track: common_track,
-        overrides: vertical_overrides,
-        fallback_color: fallback_color
-      })
+      build_axis_segments(
+        base.vertical,
+        common_thumb,
+        common_track,
+        Map.get(map, :vertical, %{}),
+        fallback_color
+      )
 
     horizontal =
-      normalize_axis_segments(%AxisSegmentsContext{
-        base: base.horizontal,
-        common_thumb: common_thumb,
-        common_track: common_track,
-        overrides: horizontal_overrides,
-        fallback_color: fallback_color
-      })
+      build_axis_segments(
+        base.horizontal,
+        common_thumb,
+        common_track,
+        Map.get(map, :horizontal, %{}),
+        fallback_color
+      )
 
     intersection =
       base.intersection
       |> merge_segment(Map.get(map, :intersection, %{}), fallback_color)
       |> segment_to_renderable()
 
-    %__MODULE__{
-      enabled: true,
-      axis: axis,
-      show: show,
-      mode: mode,
-      placement: placement,
-      vertical_placement: vertical_placement,
-      horizontal_placement: horizontal_placement,
-      min_thumb_size: min_thumb_size,
-      sizing: sizing,
-      arrows: arrows,
-      vertical: vertical,
-      horizontal: horizontal,
-      intersection: intersection
-    }
+    struct(
+      base,
+      Map.merge(map, %{
+        enabled: true,
+        vertical: vertical,
+        horizontal: horizontal,
+        intersection: intersection
+      })
+    )
   end
 
   def normalize(_other, style), do: normalize(true, style)
@@ -211,6 +210,442 @@ defmodule BackBreeze.Scrollbar do
   def segment_char(%Segment{char: char}), do: char
   def segment_char(_), do: nil
 
+  @spec add_to_layer_map(map(), map()) :: map()
+  def add_to_layer_map(layer_map, options)
+      when is_map(options) and not is_struct(options, Context) do
+    options
+    |> build_context()
+    |> then(&add_to_layer_map(layer_map, &1))
+  end
+
+  def add_to_layer_map(layer_map, %Context{} = context) do
+    cond do
+      context.style.overflow != :hidden or !context.style.scrollbar.enabled ->
+        layer_map
+
+      !is_integer(context.max_x) or !is_integer(context.max_y) ->
+        layer_map
+
+      true ->
+        context = prepare_context(context)
+
+        if !context.vertical? and !context.horizontal? do
+          layer_map
+        else
+          layer_map
+          |> maybe_clear_scrollbar_strips(context)
+          |> maybe_draw_vertical_scrollbar(context)
+          |> maybe_draw_horizontal_scrollbar(context)
+          |> maybe_draw_intersection(context)
+        end
+    end
+  end
+
+  defp build_context(%{style: style} = options) do
+    normalized = normalize(style.scrollbar, style)
+    style = %{style | scrollbar: normalized}
+    {scroll_top, scroll_left} = Map.get(options, :scroll, {0, 0})
+
+    %Context{
+      style: style,
+      content_width: Map.get(options, :content_width, 0),
+      content_height: Map.get(options, :content_height, 0),
+      scroll_top: scroll_top,
+      scroll_left: scroll_left,
+      max_x: Map.get(options, :max_x),
+      max_y: Map.get(options, :max_y)
+    }
+  end
+
+  defp prepare_context(%Context{} = context) do
+    {left, top, right, bottom} =
+      viewport_bounds(context.style.border, context.max_x, context.max_y)
+
+    viewport_width = max(right - left + 1, 0)
+    viewport_height = max(bottom - top + 1, 0)
+
+    %{
+      context
+      | left: left,
+        top: top,
+        right: right,
+        bottom: bottom,
+        viewport_width: viewport_width,
+        viewport_height: viewport_height,
+        content_width: max(context.content_width, viewport_width),
+        content_height: max(context.content_height, viewport_height)
+    }
+    |> resolve_visible_axes()
+    |> set_scrollbar_positions()
+  end
+
+  defp set_scrollbar_positions(%Context{} = context) do
+    vertical_placement = placement(context.style.scrollbar, :vertical)
+    horizontal_placement = placement(context.style.scrollbar, :horizontal)
+
+    {x_scrollbar, y_scrollbar} =
+      scrollbar_positions(context, vertical_placement, horizontal_placement)
+
+    {vertical_start, vertical_end} =
+      trim_axis_for_intersection(
+        context.top,
+        context.bottom,
+        context.horizontal? and y_scrollbar >= context.top and y_scrollbar <= context.bottom,
+        horizontal_placement
+      )
+
+    {horizontal_start, horizontal_end} =
+      trim_axis_for_intersection(
+        context.left,
+        context.right,
+        context.vertical? and x_scrollbar >= context.left and x_scrollbar <= context.right,
+        vertical_placement
+      )
+
+    %{
+      context
+      | vertical_placement: vertical_placement,
+        horizontal_placement: horizontal_placement,
+        x_scrollbar: x_scrollbar,
+        y_scrollbar: y_scrollbar,
+        vertical_start: vertical_start,
+        vertical_end: vertical_end,
+        horizontal_start: horizontal_start,
+        horizontal_end: horizontal_end
+    }
+  end
+
+  defp maybe_clear_scrollbar_strips(layer_map, context) do
+    layer_map
+    |> maybe_clear_vertical_scrollbar_strip(context)
+    |> maybe_clear_horizontal_scrollbar_strip(context)
+  end
+
+  defp maybe_clear_vertical_scrollbar_strip(
+         layer_map,
+         %Context{
+           style: %{scrollbar: %{mode: :inset}},
+           vertical?: true,
+           x_scrollbar: x_scrollbar,
+           left: left,
+           right: right,
+           vertical_start: vertical_start,
+           vertical_end: vertical_end
+         }
+       )
+       when x_scrollbar >= left and x_scrollbar <= right do
+    clear_vertical_strip(layer_map, x_scrollbar, vertical_start, vertical_end)
+  end
+
+  defp maybe_clear_vertical_scrollbar_strip(layer_map, _context), do: layer_map
+
+  defp maybe_clear_horizontal_scrollbar_strip(
+         layer_map,
+         %Context{
+           style: %{scrollbar: %{mode: :inset}},
+           horizontal?: true,
+           y_scrollbar: y_scrollbar,
+           top: top,
+           bottom: bottom,
+           horizontal_start: horizontal_start,
+           horizontal_end: horizontal_end
+         }
+       )
+       when y_scrollbar >= top and y_scrollbar <= bottom do
+    clear_horizontal_strip(layer_map, y_scrollbar, horizontal_start, horizontal_end)
+  end
+
+  defp maybe_clear_horizontal_scrollbar_strip(layer_map, _context), do: layer_map
+
+  defp viewport_bounds(border, max_x, max_y) do
+    left = if border.left, do: 1, else: 0
+    top = if border.top, do: 1, else: 0
+
+    right = max(max_x - if(border.right, do: 1, else: 0), left)
+    bottom = max(max_y - if(border.bottom, do: 1, else: 0), top)
+
+    {left, top, right, bottom}
+  end
+
+  defp scrollbar_positions(
+         %Context{
+           style: %{border: border, scrollbar: config},
+           left: left,
+           top: top,
+           right: right,
+           bottom: bottom
+         },
+         vertical_placement,
+         horizontal_placement
+       ) do
+    x_scrollbar =
+      cond do
+        config.mode == :inset and vertical_placement == :start and border.left -> left - 1
+        config.mode == :inset and vertical_placement == :end and border.right -> right + 1
+        vertical_placement == :start -> left
+        true -> right
+      end
+
+    y_scrollbar =
+      cond do
+        config.mode == :inset and horizontal_placement == :start and border.top -> top - 1
+        config.mode == :inset and horizontal_placement == :end and border.bottom -> bottom + 1
+        horizontal_placement == :start -> top
+        true -> bottom
+      end
+
+    {x_scrollbar, y_scrollbar}
+  end
+
+  defp resolve_visible_axes(%Context{} = context) do
+    config = context.style.scrollbar
+
+    vertical? =
+      axis_enabled?(config, :vertical) and
+        visible?(config, :vertical, context.content_height, context.viewport_height)
+
+    horizontal? =
+      axis_enabled?(config, :horizontal) and
+        visible?(config, :horizontal, context.content_width, context.viewport_width)
+
+    {eff_viewport_width, eff_viewport_height} =
+      effective_viewport_size(config, %{
+        width: context.viewport_width,
+        height: context.viewport_height,
+        vertical?: vertical?,
+        horizontal?: horizontal?
+      })
+
+    vertical? =
+      axis_enabled?(config, :vertical) and
+        visible?(config, :vertical, context.content_height, eff_viewport_height)
+
+    horizontal? =
+      axis_enabled?(config, :horizontal) and
+        visible?(config, :horizontal, context.content_width, eff_viewport_width)
+
+    {eff_viewport_width, eff_viewport_height} =
+      effective_viewport_size(config, %{
+        width: context.viewport_width,
+        height: context.viewport_height,
+        vertical?: vertical?,
+        horizontal?: horizontal?
+      })
+
+    %{
+      context
+      | vertical?: vertical?,
+        horizontal?: horizontal?,
+        eff_viewport_width: eff_viewport_width,
+        eff_viewport_height: eff_viewport_height
+    }
+  end
+
+  defp trim_axis_for_intersection(start_pos, end_pos, other_enabled?, other_placement) do
+    start_pos = if other_enabled? && other_placement == :start, do: start_pos + 1, else: start_pos
+    end_pos = if other_enabled? && other_placement == :end, do: end_pos - 1, else: end_pos
+
+    {start_pos, max(start_pos - 1, end_pos)}
+  end
+
+  defp clear_vertical_strip(layer_map, x, y_start, y_end) when y_start <= y_end do
+    Enum.reduce(y_start..y_end, layer_map, fn y, acc ->
+      Map.put(acc, {y, x}, {" ", ""})
+    end)
+  end
+
+  defp clear_vertical_strip(layer_map, _x, _y_start, _y_end), do: layer_map
+
+  defp clear_horizontal_strip(layer_map, y, x_start, x_end) when x_start <= x_end do
+    Enum.reduce(x_start..x_end, layer_map, fn x, acc ->
+      Map.put(acc, {y, x}, {" ", ""})
+    end)
+  end
+
+  defp clear_horizontal_strip(layer_map, _y, _x_start, _x_end), do: layer_map
+
+  defp maybe_draw_vertical_scrollbar(layer_map, %Context{vertical?: true} = context) do
+    context
+    |> vertical_axis_context()
+    |> then(&draw_axis_scrollbar(layer_map, &1))
+  end
+
+  defp maybe_draw_vertical_scrollbar(layer_map, _context), do: layer_map
+
+  defp maybe_draw_horizontal_scrollbar(layer_map, %Context{horizontal?: true} = context) do
+    context
+    |> horizontal_axis_context()
+    |> then(&draw_axis_scrollbar(layer_map, &1))
+  end
+
+  defp maybe_draw_horizontal_scrollbar(layer_map, _context), do: layer_map
+
+  defp vertical_axis_context(%Context{} = context) do
+    %AxisContext{
+      axis: :vertical,
+      config: context.style.scrollbar,
+      fixed: context.x_scrollbar,
+      start: context.vertical_start,
+      stop: context.vertical_end,
+      scroll_value: context.scroll_top,
+      content_size: context.content_height,
+      viewport_size: context.eff_viewport_height
+    }
+  end
+
+  defp horizontal_axis_context(%Context{} = context) do
+    %AxisContext{
+      axis: :horizontal,
+      config: context.style.scrollbar,
+      fixed: context.y_scrollbar,
+      start: context.horizontal_start,
+      stop: context.horizontal_end,
+      scroll_value: context.scroll_left,
+      content_size: context.content_width,
+      viewport_size: context.eff_viewport_width
+    }
+  end
+
+  defp draw_axis_scrollbar(layer_map, %AxisContext{} = axis_context) do
+    total_size = axis_context.stop - axis_context.start + 1
+
+    if total_size <= 0 do
+      layer_map
+    else
+      {track_start, track_end, layer_map} =
+        maybe_draw_axis_arrows(layer_map, axis_context, total_size)
+
+      draw_scroll_track_and_thumb(layer_map, %{
+        axis_context
+        | start: track_start,
+          stop: track_end
+      })
+    end
+  end
+
+  defp maybe_draw_axis_arrows(
+         layer_map,
+         %AxisContext{config: %{arrows: true}} = axis_context,
+         total_size
+       )
+       when total_size >= 3 do
+    layer_map =
+      put_axis_segment(
+        layer_map,
+        axis_context,
+        axis_context.start,
+        axis_arrow_segment(axis_context, :start)
+      )
+
+    layer_map =
+      put_axis_segment(
+        layer_map,
+        axis_context,
+        axis_context.stop,
+        axis_arrow_segment(axis_context, :end)
+      )
+
+    {axis_context.start + 1, axis_context.stop - 1, layer_map}
+  end
+
+  defp maybe_draw_axis_arrows(layer_map, %AxisContext{} = axis_context, _total_size) do
+    {axis_context.start, axis_context.stop, layer_map}
+  end
+
+  defp draw_scroll_track_and_thumb(
+         layer_map,
+         %AxisContext{start: track_start, stop: track_end, viewport_size: viewport_size} =
+           axis_context
+       )
+       when track_start <= track_end and viewport_size > 0 do
+    track_size = track_end - track_start + 1
+
+    {track_segment, thumb_segment} = axis_segments(axis_context)
+
+    max_scroll = max(axis_context.content_size - axis_context.viewport_size, 0)
+    scroll_value = min(max(axis_context.scroll_value, 0), max_scroll)
+
+    thumb_size =
+      thumb_size(
+        axis_context.config,
+        track_size,
+        max(axis_context.content_size, axis_context.viewport_size)
+      )
+
+    thumb_start =
+      if max_scroll == 0 or thumb_size >= track_size do
+        track_start
+      else
+        offset = round(scroll_value * (track_size - thumb_size) / max_scroll)
+        track_start + offset
+      end
+
+    thumb_end = min(thumb_start + thumb_size - 1, track_end)
+
+    layer_map =
+      Enum.reduce(track_start..track_end, layer_map, fn pos, acc ->
+        put_axis_segment(acc, axis_context, pos, track_segment)
+      end)
+
+    Enum.reduce(thumb_start..thumb_end, layer_map, fn pos, acc ->
+      put_axis_segment(acc, axis_context, pos, thumb_segment)
+    end)
+  end
+
+  defp draw_scroll_track_and_thumb(layer_map, _axis_context), do: layer_map
+
+  defp axis_segments(%AxisContext{axis: :vertical, config: config}) do
+    {config.vertical.track, config.vertical.thumb}
+  end
+
+  defp axis_segments(%AxisContext{axis: :horizontal, config: config}) do
+    {config.horizontal.track, config.horizontal.thumb}
+  end
+
+  defp axis_arrow_segment(%AxisContext{axis: :vertical, config: config}, :start),
+    do: config.vertical.arrow_start
+
+  defp axis_arrow_segment(%AxisContext{axis: :vertical, config: config}, :end),
+    do: config.vertical.arrow_end
+
+  defp axis_arrow_segment(%AxisContext{axis: :horizontal, config: config}, :start),
+    do: config.horizontal.arrow_start
+
+  defp axis_arrow_segment(%AxisContext{axis: :horizontal, config: config}, :end),
+    do: config.horizontal.arrow_end
+
+  defp put_axis_segment(layer_map, %AxisContext{} = axis_context, pos, segment) do
+    put_segment(layer_map, axis_point(axis_context, pos), segment)
+  end
+
+  defp axis_point(%AxisContext{axis: :vertical, fixed: fixed}, pos), do: {pos, fixed}
+  defp axis_point(%AxisContext{axis: :horizontal, fixed: fixed}, pos), do: {fixed, pos}
+
+  defp maybe_draw_intersection(
+         layer_map,
+         %Context{vertical?: true, horizontal?: true} = context
+       ) do
+    put_segment(
+      layer_map,
+      {context.y_scrollbar, context.x_scrollbar},
+      context.style.scrollbar.intersection
+    )
+  end
+
+  defp maybe_draw_intersection(layer_map, _context), do: layer_map
+
+  defp put_segment(layer_map, _point, nil), do: layer_map
+
+  defp put_segment(layer_map, point, segment) do
+    char = segment_char(segment)
+
+    if is_binary(char) do
+      Map.put(layer_map, point, {char, style_sequence(segment)})
+    else
+      layer_map
+    end
+  end
+
   defp default(style) do
     border_color = Map.get(style, :border_color)
 
@@ -228,66 +663,12 @@ defmodule BackBreeze.Scrollbar do
         arrow_start: %Segment{char: "◀", foreground_color: border_color},
         arrow_end: %Segment{char: "▶", foreground_color: border_color}
       },
-      intersection: %Segment{char: "┼", foreground_color: border_color}
+      intersection: %Segment{char: "┘", foreground_color: border_color}
     }
     |> renderable_segments()
   end
 
-  defp normalize_axis(value) when value in @axes, do: value
-  defp normalize_axis(_), do: :vertical
-
-  defp normalize_show(value) when value in @shows, do: value
-  defp normalize_show(_), do: :auto
-
-  defp normalize_mode(value) when value in @modes, do: value
-  defp normalize_mode(_), do: :inset
-
-  defp normalize_placement(value) when value in @placements, do: value
-  defp normalize_placement(_), do: :end
-
-  defp normalize_optional_placement(value) when value in @placements, do: value
-  defp normalize_optional_placement(_), do: nil
-
-  defp normalize_sizing(:proportional), do: :proportional
-
-  defp normalize_sizing({:fixed, value}) do
-    {:fixed, max(normalize_int(value, 1), 1)}
-  end
-
-  defp normalize_sizing(_), do: :proportional
-
-  defp normalize_boolean(value, _default) when is_boolean(value), do: value
-
-  defp normalize_boolean(value, default) when is_binary(value) do
-    case String.downcase(String.trim(value)) do
-      "true" -> true
-      "1" -> true
-      "false" -> false
-      "0" -> false
-      _ -> default
-    end
-  end
-
-  defp normalize_boolean(_value, default), do: default
-
-  defp normalize_int(value, _default) when is_integer(value), do: value
-
-  defp normalize_int(value, default) when is_binary(value) do
-    case Integer.parse(value) do
-      {value, ""} -> value
-      _ -> default
-    end
-  end
-
-  defp normalize_int(_value, default), do: default
-
-  defp normalize_axis_segments(%AxisSegmentsContext{
-         base: base,
-         common_thumb: common_thumb,
-         common_track: common_track,
-         overrides: overrides,
-         fallback_color: fallback_color
-       }) do
+  defp build_axis_segments(base, common_thumb, common_track, overrides, fallback_color) do
     overrides = if is_map(overrides), do: overrides, else: %{}
 
     thumb =
@@ -322,18 +703,13 @@ defmodule BackBreeze.Scrollbar do
     segment =
       %Segment{
         segment
-        | char: normalize_char(char, segment.char),
+        | char: char,
           foreground_color:
-            map
-            |> Map.get(:foreground_color, segment.foreground_color || fallback_color)
-            |> normalize_color(segment.foreground_color || fallback_color),
-          background_color:
-            map
-            |> Map.get(:background_color, segment.background_color)
-            |> normalize_color(segment.background_color),
-          bold: normalize_boolean(Map.get(map, :bold, segment.bold), segment.bold),
-          italic: normalize_boolean(Map.get(map, :italic, segment.italic), segment.italic),
-          reverse: normalize_boolean(Map.get(map, :reverse, segment.reverse), segment.reverse)
+            Map.get(map, :foreground_color, segment.foreground_color || fallback_color),
+          background_color: Map.get(map, :background_color, segment.background_color),
+          bold: Map.get(map, :bold, segment.bold),
+          italic: Map.get(map, :italic, segment.italic),
+          reverse: Map.get(map, :reverse, segment.reverse)
       }
 
     cond do
@@ -343,23 +719,6 @@ defmodule BackBreeze.Scrollbar do
   end
 
   defp merge_segment(%Segment{} = segment, _other, _fallback_color), do: segment
-
-  defp normalize_char(nil, default), do: default
-
-  defp normalize_char(char, _default) when is_binary(char) do
-    char
-    |> String.graphemes()
-    |> List.first()
-  end
-
-  defp normalize_char(_char, default), do: default
-
-  defp normalize_color(nil, default), do: default
-
-  defp normalize_color(value, _default) when is_integer(value) and value >= 0 and value <= 255,
-    do: value
-
-  defp normalize_color(_value, default), do: default
 
   defp renderable_segments(%__MODULE__{} = config) do
     %{
@@ -389,13 +748,10 @@ defmodule BackBreeze.Scrollbar do
       |> maybe_style(:italic, segment.italic)
       |> maybe_style(:reverse, segment.reverse)
 
-    token = "§"
-
     style_prefix =
       style
-      |> Termite.Style.render_to_string(token)
-      |> String.split(token, parts: 2)
-      |> List.first()
+      |> Termite.Style.render_to_string("")
+      |> String.trim_trailing(Termite.Style.reset_code())
 
     %{segment | style: style_prefix}
   end
